CprAnal: Activate Sum CPR for local bindings
We've had Sum CPR (#5075) for top-level bindings for a couple of years
now. That begs the question why we didn't also activate it for local
bindings, and the reasons for that are described in
`Note [CPR for sum types]`. Only that it didn't make sense! The Note
said that Sum CPR would destroy let-no-escapes, but that should be a
non-issue since we have syntactic join points in Core now and we don't
WW for them (`Note [Don't w/w join points for CPR]`).

Doing more WW had the consequence that `eftInt` and `eftWord` won't
inline in some cases, so I marked it `INLINABLE [1]`. Similarly
`efdtInt` and `efdtWord`.

So I simply activated CPR for all bindings of sum type, thus
fixing #5075.